+++ src/Task_6.java
@@ -1,44 +1,36 @@
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.locks.ReentrantReadWriteLock;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.atomic.AtomicInteger;
 
 public class Task_6 {
 
-    public final List<String> notes = new ArrayList<>();
-    private final ReentrantReadWriteLock lock = new ReentrantReadWriteLock();
 
-    public void addNote(int index, String note) {
-        System.out.println("Сейчас будет добавлена заметка [" + note + "] На позицию " + index);
-        lock.writeLock().lock();
-        try {
-            notes.add(index, note);
-        } finally {
-            lock.writeLock().unlock();
+    public static void main(String[] args) {
+        TreasureChest chest = new TreasureChest();
+        ExecutorService executor = Executors.newFixedThreadPool(5);
+
+        for (int i = 0; i < 5; i++) {
+            executor.execute(() -> {
+                chest.add(50);
+            });
         }
-        System.out.println("Уже добавлена заметка [" + note + "]");
+
+        executor.shutdown();
+        while (!executor.isTerminated()) {
+        }
+
+        System.out.println("Общее количество золота в сундуке: " + chest.getTotal());
+    }
+}
+
+class TreasureChest {
+    private final AtomicInteger goldCount = new AtomicInteger(0);
+
+    public void add(int amount) {
+        goldCount.addAndGet(amount);
     }
 
-    public void removeNote(int index) {
-        System.out.println("Сейчас будет удалена заметка с позиции " + index);
-        String note;
-        lock.writeLock().lock();
-        try {
-            note = notes.remove(index);
-        } finally {
-            lock.writeLock().unlock();
-        }
-        System.out.println("Уже удалена заметка [" + note + "] с позиции " + index);
-    }
-
-    public void readNote(int index) {
-        System.out.println("Сейчас будет прочтена заметка с позиции " + index);
-        String note;
-        lock.readLock().lock();
-        try {
-            note = notes.get(index);
-        } finally {
-            lock.readLock().unlock();
-        }
-        System.out.println("Прочтена заметка [" + note + "] с позиции " + index);
+    public int getTotal() {
+        return goldCount.get();
     }
 }
